Renamed modal files (#1066)
